renaming configuration classes
